+++ app/src/main/java/com/example/twitterminiapp/presentation/di/application/TwitterApp.kt
@@ -1,4 +1,4 @@
-package com.example.twitterminiapp.presentation.di
+package com.example.twitterminiapp.presentation.di.application
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
