+++ app/src/main/java/com/example/twitterminiapp/presentation/di/modules/RepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.example.twitterminiapp.presentation.di
+package com.example.twitterminiapp.presentation.di.modules
 
 import com.example.twitterminiapp.data.repository.TwitterRepositoryImpl
 import com.example.twitterminiapp.data.repository.dataSource.TwitterRemoteDataSource
