+++ app/src/main/java/com/example/twitterminiapp/presentation/fragment/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.twitterminiapp
+package com.example.twitterminiapp.presentation.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
