+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/profile_background"
-    tools:context=".ProfileFragment">
+    tools:context=".presentation.fragment.ProfileFragment">
 
     <ImageView
         android:id="@+id/user_icon"
